Add Working W3mImgDisplay & Brick Widget Modules With Example App
This adds a `W3mImgDisplay` module which is responsible for calling
`w3mimgdisplay` with specific commands.

The `Brick.Widgets.Image` module contains an `ImageWidget` & related
functions for rendering images directly into Brick applications.

The example application shows a list of images in the current directory
along with a preview of the currently selected image.